+++ pkg/controllers/requeableErrors_test.go
@@ -3,7 +3,6 @@ package controllers
 import (
 	"errors"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -30,27 +29,4 @@ func Test_genericRequeueableError_Error(t *testing.T) {
 	sut := &genericRequeueableError{"oh noez", assert.AnError}
 	expected := "oh noez: " + assert.AnError.Error()
 	assert.Equal(t, expected, sut.Error())
-}
-
-func Test_genericRequeueableError_GetRequeueTime(t *testing.T) {
-	type args struct {
-		requeueTime        time.Duration
-		defaultRequeueTime time.Duration
-	}
-	tests := []struct {
-		name string
-		args args
-		want time.Duration
-	}{
-		{"always return defaultRequeueTime case 1", args{0 * time.Second, 5 * time.Second}, 5 * time.Second},
-		{"always return defaultRequeueTime case 2", args{15 * time.Second, 5 * time.Second}, 5 * time.Second},
-		{"always return defaultRequeueTime case 3", args{30 * time.Second, 5 * time.Second}, 5 * time.Second},
-		{"always return defaultRequeueTime case 3", args{256 * time.Second, 5 * time.Second}, 5 * time.Second},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			due := &genericRequeueableError{}
-			assert.Equalf(t, tt.want, due.GetRequeueTime(tt.args.requeueTime, tt.args.defaultRequeueTime), "getRequeueTime(%v)", tt.args.requeueTime)
-		})
-	}
 }