+++ pkg/helm/mock_HelmClient_test.go
@@ -28,9 +28,9 @@ func (_m *MockHelmClient) EXPECT() *MockHelmClient_Expecter {
 	return &MockHelmClient_Expecter{mock: &_m.Mock}
 }
 
-// GetChart provides a mock function with given fields: spec
-func (_m *MockHelmClient) GetChart(spec *client.ChartSpec) (*chart.Chart, string, error) {
-	ret := _m.Called(spec)
+// GetChart provides a mock function with given fields: ctx, spec
+func (_m *MockHelmClient) GetChart(ctx context.Context, spec *client.ChartSpec) (*chart.Chart, string, error) {
+	ret := _m.Called(ctx, spec)
 
 	if len(ret) == 0 {
 		panic("no return value specified for GetChart")
@@ -39,25 +39,25 @@ func (_m *MockHelmClient) GetChart(spec *client.ChartSpec) (*chart.Chart, string
 	var r0 *chart.Chart
 	var r1 string
 	var r2 error
-	if rf, ok := ret.Get(0).(func(*client.ChartSpec) (*chart.Chart, string, error)); ok {
-		return rf(spec)
+	if rf, ok := ret.Get(0).(func(context.Context, *client.ChartSpec) (*chart.Chart, string, error)); ok {
+		return rf(ctx, spec)
 	}
-	if rf, ok := ret.Get(0).(func(*client.ChartSpec) *chart.Chart); ok {
-		r0 = rf(spec)
+	if rf, ok := ret.Get(0).(func(context.Context, *client.ChartSpec) *chart.Chart); ok {
+		r0 = rf(ctx, spec)
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).(*chart.Chart)
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(*client.ChartSpec) string); ok {
-		r1 = rf(spec)
+	if rf, ok := ret.Get(1).(func(context.Context, *client.ChartSpec) string); ok {
+		r1 = rf(ctx, spec)
 	} else {
 		r1 = ret.Get(1).(string)
 	}
 
-	if rf, ok := ret.Get(2).(func(*client.ChartSpec) error); ok {
-		r2 = rf(spec)
+	if rf, ok := ret.Get(2).(func(context.Context, *client.ChartSpec) error); ok {
+		r2 = rf(ctx, spec)
 	} else {
 		r2 = ret.Error(2)
 	}
@@ -71,14 +71,15 @@ type MockHelmClient_GetChart_Call struct {
 }
 
 // GetChart is a helper method to define mock.On call
+//   - ctx context.Context
 //   - spec *client.ChartSpec
-func (_e *MockHelmClient_Expecter) GetChart(spec interface{}) *MockHelmClient_GetChart_Call {
-	return &MockHelmClient_GetChart_Call{Call: _e.mock.On("GetChart", spec)}
+func (_e *MockHelmClient_Expecter) GetChart(ctx interface{}, spec interface{}) *MockHelmClient_GetChart_Call {
+	return &MockHelmClient_GetChart_Call{Call: _e.mock.On("GetChart", ctx, spec)}
 }
 
-func (_c *MockHelmClient_GetChart_Call) Run(run func(spec *client.ChartSpec)) *MockHelmClient_GetChart_Call {
+func (_c *MockHelmClient_GetChart_Call) Run(run func(ctx context.Context, spec *client.ChartSpec)) *MockHelmClient_GetChart_Call {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(*client.ChartSpec))
+		run(args[0].(context.Context), args[1].(*client.ChartSpec))
 	})
 	return _c
 }
@@ -88,7 +89,7 @@ func (_c *MockHelmClient_GetChart_Call) Return(_a0 *chart.Chart, _a1 string, _a2
 	return _c
 }
 
-func (_c *MockHelmClient_GetChart_Call) RunAndReturn(run func(*client.ChartSpec) (*chart.Chart, string, error)) *MockHelmClient_GetChart_Call {
+func (_c *MockHelmClient_GetChart_Call) RunAndReturn(run func(context.Context, *client.ChartSpec) (*chart.Chart, string, error)) *MockHelmClient_GetChart_Call {
 	_c.Call.Return(run)
 	return _c
 }